Extract logic from `VariedSubject` fragments
