+++ app/src/main/java/com/senex/timetable/presentation/ui/subject/varied/english/NewEnglishSubjectFragment.kt
@@ -1,53 +1,29 @@
 package com.senex.timetable.presentation.ui.subject.varied.english
 
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.lifecycle.lifecycleScope
-import androidx.navigation.fragment.findNavController
+import androidx.navigation.NavDirections
 import androidx.navigation.fragment.navArgs
-import com.senex.timetable.databinding.FragmentVariedSubjectBinding
+import com.senex.timetable.domain.model.subject.EnglishSubject
 import com.senex.timetable.presentation.common.assistedViewModel
-import com.senex.timetable.presentation.common.initNavToolbar
-import com.senex.timetable.presentation.ui.subject.common.initShowHideSubjectButtons
-import com.senex.timetable.presentation.ui.subject.varied.BindingFragment
+import com.senex.timetable.presentation.ui.subject.varied.VariedSubjectFragment
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-class NewEnglishSubjectFragment : BindingFragment<FragmentVariedSubjectBinding>() {
-
+class NewEnglishSubjectFragment : VariedSubjectFragment<EnglishSubject>() {
     private val args: NewEnglishSubjectFragmentArgs by navArgs()
 
     @Inject
     lateinit var factory: NewEnglishSubjectViewModel.Factory
-    private val viewModel: NewEnglishSubjectViewModel by assistedViewModel {
+    override val viewModel: NewEnglishSubjectViewModel by assistedViewModel {
         factory.create(args.englishSubjectId)
     }
 
-    override val bindingInflater: (LayoutInflater, ViewGroup?, Boolean) -> FragmentVariedSubjectBinding =
-        FragmentVariedSubjectBinding::inflate
-
-    private val navigateToSelectionFragment: (View) -> Unit = {
-        lifecycleScope.launch {
-            findNavController().navigate(
-                NewEnglishSubjectFragmentDirections.actionNewEnglishSubjectFragmentToNewSelectableEnglishSubjectsFragment(
-                    args.englishSubjectId,
-                    viewModel.variedSubject.first().primarySubjectId ?: -1,
-                )
-            )
-        }
-    }
-
-    override val onViewCreatedCallback: FragmentVariedSubjectBinding.() -> Unit = {
-        subjectShowHideButtons.initShowHideSubjectButtons(
-            viewLifecycleOwner.lifecycleScope,
-            viewModel.isVariedSubjectVisible,
-            viewModel::setSubjectVisibility,
+    override val selectionFragmentNavDirections: suspend () -> NavDirections = {
+        NewEnglishSubjectFragmentDirections.actionNewEnglishSubjectFragmentToNewSelectableEnglishSubjectsFragment(
+            args.englishSubjectId,
+            viewModel.variedSubject.first().primarySubjectId ?: -1,
         )
-        chooseCourseButton.setOnClickListener(navigateToSelectionFragment)
-        toolbarContainer.toolbar.initNavToolbar(findNavController())
     }
+
 }
